+++ module-services/service-audio/ServiceAudio.cpp
@@ -73,9 +73,10 @@ constexpr bool ServiceAudio::IsResumable(const audio::PlaybackType &type) const
     return type == audio::PlaybackType::Multimedia;
 }
 
-constexpr bool ServiceAudio::ShouldLoop(const audio::PlaybackType &type) const
+constexpr bool ServiceAudio::ShouldLoop(const std::optional<audio::PlaybackType> &type) const
 {
-    return type == audio::PlaybackType::CallRingtone;
+
+    return type.value_or(audio::PlaybackType::None) == audio::PlaybackType::CallRingtone;
 }
 
 std::unique_ptr<AudioResponseMessage> ServiceAudio::HandlePause(std::optional<AudioRequestMessage *> msg)
@@ -85,12 +86,13 @@ std::unique_ptr<AudioResponseMessage> ServiceAudio::HandlePause(std::optional<Au
     audio::Token retToken  = Token::MakeBadToken();
 
     auto pauseInput = [this](AudioMux::Input &audioInput) {
-        if (IsResumable(audioInput.audio.GetCurrentOperation()->GetPlaybackType())) {
-            audioInput.audio.Pause();
+        auto playbackType = audioInput.audio->GetCurrentOperationPlaybackType();
+        if (playbackType && IsResumable(*playbackType)) {
+            audioInput.audio->Pause();
             return audioInput.token;
         }
         else {
-            audioInput.audio.Stop();
+            audioInput.audio->Stop();
             auto broadMsg = std::make_shared<AudioNotificationMessage>(
                 static_cast<AudioNotificationMessage::Type>(AudioNotificationMessage::Type::Stop), audioInput.token);
             sys::Bus::SendMulticast(broadMsg, sys::BusChannels::ServiceAudioNotifications, this);
@@ -107,7 +109,7 @@ std::unique_ptr<AudioResponseMessage> ServiceAudio::HandlePause(std::optional<Au
     }
     else {
         for (auto &audioInput : audioMux.GetAllInputs()) {
-            if (audioInput.audio.GetCurrentOperation()->GetState() == Operation::State::Active) {
+            if (audioInput.audio->GetCurrentOperation()->GetState() == Operation::State::Active) {
                 pauseInput(audioInput);
             }
         }
@@ -130,7 +132,7 @@ std::unique_ptr<AudioResponseMessage> ServiceAudio::HandleStart(std::optional<Au
 
         HandlePause();
         retToken = audioMux.IncrementToken(input);
-        retCode  = (*input)->audio.Start(opType, retToken, args...);
+        retCode  = (*input)->audio->Start(opType, retToken, args...);
     }
     if (retCode != audio::RetCode::Success) {
         retToken = Token::MakeBadToken();
@@ -144,10 +146,10 @@ std::unique_ptr<AudioResponseMessage> ServiceAudio::HandleStop(AudioStopMessage 
     std::vector<audio::RetCode> retCodes = {audio::RetCode::Success};
 
     auto stopInput = [this](auto inp) {
-        if (inp->audio.GetCurrentState() == Audio::State::Idle) {
+        if (inp->audio->GetCurrentState() == Audio::State::Idle) {
             return audio::RetCode::Success;
         }
-        auto rCode = inp->audio.Stop();
+        auto rCode = inp->audio->Stop();
         // Send notification that audio file was stopped
         auto msgStop = std::make_shared<AudioNotificationMessage>(
             static_cast<AudioNotificationMessage::Type>(AudioNotificationMessage::Type::Stop), inp->token);
@@ -163,7 +165,7 @@ std::unique_ptr<AudioResponseMessage> ServiceAudio::HandleStop(AudioStopMessage 
     // stop with vector of playback types
     else if (!playbackTypesToStop.empty()) {
         for (auto &input : audioMux.GetAllInputs()) {
-            const auto currentOperation = input.audio.GetCurrentOperation();
+            const auto currentOperation = input.audio->GetCurrentOperation();
             if (currentOperation && (std::find(playbackTypesToStop.begin(),
                                                playbackTypesToStop.end(),
                                                currentOperation->GetPlaybackType()) != playbackTypesToStop.end())) {
@@ -198,8 +200,8 @@ sys::Message_t ServiceAudio::DataReceivedHandler(sys::DataMessage *msgl, sys::Re
 
         case AudioNotificationMessage::Type::EndOfFile: {
             auto input = audioMux.GetInput(msg->token);
-            if (input && ShouldLoop((*input)->audio.GetCurrentOperation()->GetPlaybackType())) {
-                (*input)->audio.Start();
+            if (input && ShouldLoop((*input)->audio->GetCurrentOperationPlaybackType())) {
+                (*input)->audio->Start();
             }
             else {
                 auto newMsg = std::make_shared<AudioStopMessage>(msg->token);
@@ -238,7 +240,7 @@ sys::Message_t ServiceAudio::DataReceivedHandler(sys::DataMessage *msgl, sys::Re
         case MessageType::AudioRecorderStart: {
             if (auto input = audioMux.GetRecordingInput()) {
                 responseMsg = std::make_shared<AudioResponseMessage>(
-                    (*input)->audio.Start(Operation::Type::Recorder, (*input)->token, msg->fileName.c_str()));
+                    (*input)->audio->Start(Operation::Type::Recorder, (*input)->token, msg->fileName.c_str()));
             }
         } break;
 
@@ -255,13 +257,13 @@ sys::Message_t ServiceAudio::DataReceivedHandler(sys::DataMessage *msgl, sys::Re
 
         case MessageType::AudioResume: {
             if (auto input = audioMux.GetInput(msg->token)) {
-                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio.Resume());
+                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio->Resume());
             }
         } break;
 
         case MessageType::AudioGetFileTags: {
-            if (auto input = audioMux.GetInput()) {
-                auto tag = (*input)->audio.GetFileTags(msg->fileName.c_str());
+            if (auto input = audioMux.GetAvailableInput()) {
+                auto tag = (*input)->audio->GetFileTags(msg->fileName.c_str());
                 if (tag) {
                     responseMsg = std::make_shared<AudioResponseMessage>(RetCode::Success, tag.value());
                 }
@@ -273,28 +275,27 @@ sys::Message_t ServiceAudio::DataReceivedHandler(sys::DataMessage *msgl, sys::Re
 
         case MessageType::AudioRoutingRecordCtrl: {
             if (auto input = audioMux.GetRecordingInput()) {
-                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio.SendEvent(
+                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio->SendEvent(
                     msg->enable ? Operation::Event::StartCallRecording : Operation::Event::StopCallRecording));
             }
         } break;
 
         case MessageType::AudioRoutingMute: {
             if (auto input = audioMux.GetRecordingInput()) {
-                responseMsg = std::make_shared<AudioResponseMessage>(
-                    (*input)->audio.SendEvent(msg->enable ? Operation::Event::CallMute : Operation::Event::CallUnmute));
+                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio->SendEvent(
+                    msg->enable ? Operation::Event::CallMute : Operation::Event::CallUnmute));
             }
         } break;
 
         case MessageType::AudioRoutingSpeakerhone: {
             if (auto input = audioMux.GetRecordingInput()) {
-                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio.SendEvent(
+                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio->SendEvent(
                     msg->enable ? Operation::Event::CallSpeakerphoneOn : Operation::Event::CallSpeakerphoneOff));
             }
         } break;
-
         case MessageType::AudioRoutingHeadset: {
             if (auto input = audioMux.GetRecordingInput()) {
-                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio.SendEvent(
+                responseMsg = std::make_shared<AudioResponseMessage>((*input)->audio->SendEvent(
                     msg->enable ? Operation::Event::HeadphonesPlugin : Operation::Event::HeadphonesUnplug));
             }
         } break;
@@ -372,14 +373,15 @@ std::optional<uint32_t> ServiceAudio::getCurrentSetting(const audio::Setting &se
     if (!activeInput.has_value()) {
         const auto idleInput = audioMux.GetIdleInput();
         if (idleInput.has_value() && setting == audio::Setting::Volume) {
-            const auto path = audio::str(
-                audio::PlaybackType::CallRingtone, audio::Setting::Volume, (*idleInput)->audio.GetHeadphonesInserted());
+            const auto path = audio::str(audio::PlaybackType::CallRingtone,
+                                         audio::Setting::Volume,
+                                         (*idleInput)->audio->GetHeadphonesInserted());
             return fetchAudioSettingFromDb<audio::Volume>(path, 0);
         }
         return {};
     }
 
-    const auto currentOperation = (*activeInput)->audio.GetCurrentOperation();
+    const auto currentOperation = (*activeInput)->audio->GetCurrentOperation();
     const auto path =
         audio::str(currentOperation->GetProfile()->GetType(), setting, currentOperation->GetPlaybackType());
     return (setting == audio::Setting::Volume) ? fetchAudioSettingFromDb<audio::Volume>(path, 0)
@@ -393,8 +395,9 @@ void ServiceAudio::setCurrentSetting(const audio::Setting &setting, const uint32
     if (!activeInput.has_value()) {
         const auto idleInput = audioMux.GetIdleInput();
         if (idleInput.has_value() && setting == audio::Setting::Volume) {
-            const auto path = audio::str(
-                audio::PlaybackType::CallRingtone, audio::Setting::Volume, (*idleInput)->audio.GetHeadphonesInserted());
+            const auto path = audio::str(audio::PlaybackType::CallRingtone,
+                                         audio::Setting::Volume,
+                                         (*idleInput)->audio->GetHeadphonesInserted());
             addOrIgnoreEntry(path, audio::playbackDefaults::defaultLoudspeakerVolume);
             const auto valueToSet = std::clamp(value, minVolume, maxVolume);
             updateDbValue(path, Setting::Volume, valueToSet);
@@ -403,16 +406,16 @@ void ServiceAudio::setCurrentSetting(const audio::Setting &setting, const uint32
     }
 
     auto &audio                 = (*activeInput)->audio;
-    const auto currentOperation = audio.GetCurrentOperation();
+    const auto currentOperation = audio->GetCurrentOperation();
     switch (setting) {
     case audio::Setting::Volume: {
         const auto valueToSet = std::clamp(value, minVolume, maxVolume);
-        audio.SetOutputVolume(valueToSet);
+        audio->SetOutputVolume(valueToSet);
         updateDbValue(currentOperation, audio::Setting::Volume, valueToSet);
     } break;
     case audio::Setting::Gain: {
         const auto valueToSet = std::clamp(value, minGain, maxGain);
-        audio.SetInputGain(valueToSet);
+        audio->SetInputGain(valueToSet);
         updateDbValue(currentOperation, audio::Setting::Gain, valueToSet);
     } break;
     }